requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
PyMySQL==1.0.2
cryptography==3.4.8
python-dotenv==0.19.1
Werkzeug==2.0.1
bcrypt==3.2.0
marshmallow==3.14.0